Revert "563 20 beta federation service breaks after a while"
